services/authgw/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.0
pyyaml==6.0.2
redis==5.0.8
prometheus-client>=0.20.0
